Refstructured core models and addedbgeneric invitation serializer and extra validations
