+++ src/core/models/__init__.py
@@ -1,0 +1,2 @@
+from .invitation import BaseInvitation
+from .notification import Notification
