+++ src/core/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .invitation import BaseInvitationSerializer
